refactor: extract abyss http feature into ferriskey-api-abyss
