+++ api/src/application/http/abyss.rs
@@ -1,0 +1,1 @@
+pub use ferriskey_api_abyss::*;
