+++ libs/ferriskey-api-abyss/src/federation/handlers/create_provider.rs
@@ -13,16 +13,12 @@ use ferriskey_core::domain::{
 use std::str::FromStr;
 use uuid::Uuid;
 
-use crate::application::http::{
-    abyss::federation::dto::{CreateProviderRequest, ProviderResponse},
-    server::{
-        api_entities::{
-            api_error::{ApiError, ApiErrorResponse},
-            response::Response,
-        },
-        app_state::AppState,
-    },
+use crate::federation::dto::{CreateProviderRequest, ProviderResponse};
+use ferriskey_api_core::api_entities::{
+    api_error::{ApiError, ApiErrorResponse},
+    response::Response,
 };
+use ferriskey_api_core::app_state::AppState;
 
 #[utoipa::path(
     post,
