+++ libs/ferriskey-api-abyss/src/federation/handlers/delete_provider.rs
@@ -10,13 +10,11 @@ use tracing::info;
 use utoipa::ToSchema;
 use uuid::Uuid;
 
-use crate::application::http::server::{
-    api_entities::{
-        api_error::{ApiError, ApiErrorResponse},
-        response::Response,
-    },
-    app_state::AppState,
+use ferriskey_api_core::api_entities::{
+    api_error::{ApiError, ApiErrorResponse},
+    response::Response,
 };
+use ferriskey_api_core::app_state::AppState;
 
 #[derive(Serialize, PartialEq, ToSchema)]
 pub struct DeleteProviderResponse {
