+++ libs/ferriskey-api-abyss/src/federation/handlers/get_provider.rs
@@ -7,16 +7,12 @@ use ferriskey_core::domain::{
 };
 use uuid::Uuid;
 
-use crate::application::http::{
-    abyss::federation::dto::ProviderResponse,
-    server::{
-        api_entities::{
-            api_error::{ApiError, ApiErrorResponse},
-            response::Response,
-        },
-        app_state::AppState,
-    },
+use crate::federation::dto::ProviderResponse;
+use ferriskey_api_core::api_entities::{
+    api_error::{ApiError, ApiErrorResponse},
+    response::Response,
 };
+use ferriskey_api_core::app_state::AppState;
 
 #[utoipa::path(
     get,
