+++ libs/ferriskey-api-abyss/src/identity_provider/handlers/delete_identity_provider.rs
@@ -1,14 +1,10 @@
-use crate::application::http::{
-    abyss::identity_provider::dto::DeleteIdentityProviderResponse,
-    server::{
-        api_entities::{api_error::ApiError, response::Response},
-        app_state::AppState,
-    },
-};
+use crate::identity_provider::dto::DeleteIdentityProviderResponse;
 use axum::{
     Extension,
     extract::{Path, State},
 };
+use ferriskey_api_core::api_entities::{api_error::ApiError, response::Response};
+use ferriskey_api_core::app_state::AppState;
 use ferriskey_core::domain::abyss::identity_provider::{
     entities::DeleteIdentityProviderInput, ports::IdentityProviderService,
 };
