+++ libs/ferriskey-api-abyss/src/routes.rs
@@ -1,11 +1,7 @@
-use crate::application::{
-    auth::auth,
-    http::{
-        abyss::{federation::federation_routes, identity_provider::identity_provider_routes},
-        server::app_state::AppState,
-    },
-};
+use crate::{federation::federation_routes, identity_provider::identity_provider_routes};
 use axum::{Router, middleware};
+use ferriskey_api_core::app_state::AppState;
+use ferriskey_api_core::auth::auth;
 
 pub fn abyss_routes(state: AppState) -> Router<AppState> {
     Router::new()
